+++ Engine/transport.cpp
@@ -1,26 +1,37 @@
 #include "transport.hh"
 #include "hex.hh"
 #include <memory>
+#include <algorithm>
 
-namespace Common {
+namespace Common{
 
-Transport::Transport( int transportId ): id_(transportId) {}
+Transport::Transport( int id ):
+    id_(id){}
+
 Transport::~Transport(){}
 
-void Transport::addPawn( int pawnId ) {}
-
-void Transport::move(std::shared_ptr<Hex> to)
+void Transport::addPawn(std::shared_ptr<Pawn> pawn )
 {
-    addHex(to);
+    if ( getCapacity() > 0 ){
+        pawns_.push_back(pawn);
+    }
 }
 
-int Transport::getCapacity() const{}
-
-bool Transport::canMove( int playerId ) const {}
+int Transport::getCapacity() const
+{
+    return 1234;
+}
 
 void Transport::addHex( std::shared_ptr<Common::Hex> hex )
 {
     hex->addTransport(shared_from_this());
+    hex->removeTransport(shared_from_this());
+    hex_ = hex;
+}
+
+bool Transport::isPawnInTransport(std::shared_ptr<Pawn> pawn)
+{
+    return (std::find(pawns_.begin(),pawns_.end(),pawn) != pawns_.end());
 }
 
 int Transport::getId()