+++ CHANGELOG.md
@@ -4,12 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0]
+### Changed
+- IGameBoard has been expanded, update your implementation.
+
+### Added
+- Support for specifying own spinner layout with Assets/layout.json
+- 2 new Transports, Boat and Dolphin
+- 4 new Actors, Shark, Kraken, Seamunster and Vortex
+- Added new functions to IGameBoard, addActor, addTransport, moveTransport and removeTransport. Update your own GameBoard implementation!
+
 ## [2.0.0]
 ### Changed
 - Renamed Common::Vehicle to Common::Transport
 - Wrapped Common::getGameRunner (and the whole initialization module) to namespace Initialization
 - Common::Initialization::getGameRunner now handles players as std::vector<std::shared_ptr<Common::IPlayer>> instead of std::vector<Common::IPlayer*>
-
+- 
 
 ### Added
 - Added new services to IGameEngine.